changed a bunch

+++ database/storage.go
@@ -1,6 +1,7 @@
 package database
 
 import (
+	"bytes"
 	"fmt"
 	"image"
 	"image/jpeg"
@@ -13,27 +14,30 @@ import (
 )
 
 type StorageAPI interface {
-	SaveImage(image image.Image, model Model) (string, error)
+	SaveImage(model Model) (string, error)
 	LoadImage(path string) (image.Image, string, error)
 }
 
-const imageRoot = "/images/"
+const ImageRoot = "/images/"
 
-//SaveImage takes the image and saves it to the disk at the path
-func (instance *instance) SaveImage(image image.Image, model Model) (string, error) {
-	path := imageRoot + fmt.Sprintf("%d", model.UID) + "-" + model.FileName
-	// if ok := fileExists(path); ok {
+//SaveImage takes the image and saves it to the disk and returns at the path
+func (instance *instance) SaveImage(model Model) (string, error) {
+	path := ImageRoot + fmt.Sprintf("%d", model.UID) + "-" + model.FileName
+	if ok := fileExists(path); ok {
+		//file exists we a;ready have it saved
+		return path, nil
+	}
 	file, err := os.Create(path)
 	if err != nil {
 		l.Error(errors.Wrap(err, "Error saving file: "+fmt.Sprintf("%d", model.UID)))
 		return "", err
 	}
 	defer file.Close()
-	return path, jpeg.Encode(file, image, &jpeg.Options{95}) //default
-
-	// } else {
-	// 	return
-	// }
+	img, _, err := image.Decode(bytes.NewReader(model.Image)) //forward
+	if err != nil {
+		return path, err
+	}
+	return path, jpeg.Encode(file, img, &jpeg.Options{95}) //default
 
 }
